requirements.txt
aniso8601==9.0.1
apispec==6.6.1
Authlib==1.3.0
blinker==1.8.1
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
flask-restful-swagger-3==0.5.1
flask-smorest==0.44.0
flask-swagger-ui==4.11.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
oauthlib==2.1.0
packaging==24.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.1.0
six==1.16.0
urllib3==2.2.1
webargs==8.4.0
Werkzeug==2.2.2
